[ir_rf_base] Share the entity, call and API reply code between infrared and radio_frequency
