+++ esphome/components/remote_base/remote_base.cpp
@@ -195,7 +195,7 @@ void RemoteTransmitterBase::send_(uint32_t send_times, uint32_t send_wait, [[may
     ESP_LOGVV(TAG, "%s", buffer);
   }
 #endif
-#if defined(USE_IR_RF) || defined(USE_RADIO_FREQUENCY)
+#ifdef USE_IR_RF
   this->current_seq_ = seq;
 #endif
   this->send_internal(send_times, send_wait);
